feat: Option for bottom positioning of progress bars.
BREAKING CHANGE: Though not technically breaking (defaults behave as previously), added the option for bottom positioning of progress bars. As a result, the logic is easier. Pass 'bottom' to options.anchor in the constructor to switch. There is also now a options.persist setting. If set to false or left blank, the virtual console will reconnect the original console after all tasks are finished. However, if you call updateTask or similar, it will re-intercept console. If persist is true, it will not reconnect original; intended for use with a perpetual watcher.